Add files via upload
The project revolves around analyzing and visualizing sales data for a fictional e-commerce company. The goal is to provide insights into the company's sales performance, customer behavior, and product trends. The project incorporates both Power BI and SQL to gather, clean, transform, and present the data in an interactive dashboard.